Fixed a race condition
In "handleUserAuthorisedAndEnrolled" in the base controller, we cache some
data then redirect the user to a page which depends on that data being
present in the cache.

As caching is async, we are vulnerable to the redirection being processed
before the caching has completed - therefore getting an error when the
new page tries to retrieve data which hasn't yet been cached.

The code has now been updated such that the redirection only occurs once
the caching has completed, removing this race condition.